Server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["NikshetSteh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
bcrypt = "^4.1.3"
pydantic-settings = "^2.3.4"
fastapi-pagination = "^0.12.26"
insightface = "^0.7.3"
chromadb = "^0.5.4"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
uvicorn = "^0.30.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
